Transformation Layer Implementation V1
Current implementation presents a minimally sophisticated
process of cleaning and formatting the data retrieved from the
extraction layer.

The end result is a dict[str: pd.DataFrame] which will go to the
Load layer.

Future updates to the transformation layer will check for
data integrity and potentially other additional functions
as the project unfolds.

The transformation layer taught me a lot about how
versatile and flexible data transformation can be. This process
requires a level of innovation and planning to create data
structures and transform raw data into something that can be
usable.

This layer follows a dynamic approach, multiple stocks can be
extracted and transformed in a single prompt which is more
sophisticated than seperately prompting each stock for its data.

The dynamic approach proposed many challenges which broaden and
forced me to critically think through how store all this data
in a format that practical and eifficient.

Overall, this layer so far is the most challenging and the next phase
will go into the loading layer.