Transformation Layer Implementation 1.5.1 Update
The transformation layer will feature one more process,
this process will store the data into a list of tuples where
each tuple is one stock price, this will enable a more efficient
and practical loading of data into the database.

This process is necessary for the loading phase, where the
transformed data will be loaded in batches that follows the
stock_price table design.

This layer is even more sophisticated but more intuitive in the aspect
that the data can be manipulated anytime into any reasonable format.